+++ app/src/main/java/com/ichi2/apisample/ui/OnFieldRadioChangeListener.java
@@ -1,0 +1,17 @@
+package com.ichi2.apisample.ui;
+
+import android.widget.RadioGroup;
+
+public class OnFieldRadioChangeListener implements RadioGroup.OnCheckedChangeListener {
+    private MainActivity mainActivity;
+
+    public OnFieldRadioChangeListener(MainActivity mainActivity) {
+        this.mainActivity = mainActivity;
+    }
+
+    @Override
+    public void onCheckedChanged(RadioGroup radioGroup, int i) {
+        mainActivity.clearAddedFilenames();
+        mainActivity.refreshExisting();
+    }
+}
